requirements.txt
Flask==2.2.5
argon2-cffi==23.1.0
cryptography==41.0.2
